requirements.txt
google-genai
pandas
streamlit
python-dotenv
plotly